+++ src/Message.php
@@ -2,7 +2,7 @@
 /**
  * This file is part of the "litgroup/sms" package.
  *
- * (c) LitGroup <http://litgroup.ru/>
+ * (c) Roman Shamritskiy <roman@litgroup.ru>
  *
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
@@ -10,10 +10,14 @@
 
 namespace LitGroup\Sms;
 
+use LitGroup\Sms\Exception\InvalidArgumentException;
+
 /**
  * Message.
  *
  * @author Roman Shamritskiy <roman@litgroup.ru>
+ *
+ * @api
  */
 class Message
 {
@@ -37,31 +41,16 @@ class Message
      * Message constructor.
      *
      * @param string $body Text of message (UTF-8 string).
-     * @param array $recipients List of phone numbers of recipients. Phone number should be in the format
-     *                          "+7000123456789" or "7000123456789".
-     * @param null $sender      Identifier of sender. (For example "MyCompany"). Identifier should be allowed by
-     *                          SMS-Gateway provider.
+     * @param string[] $recipients List of phone numbers of recipients. Phone number should be in the format
+     *                             "+7000123456789" or "7000123456789".
+     * @param null $sender Identifier of sender. (For example "MyCompany"). Identifier should be allowed by
+     *                     SMS-Gateway provider.
      */
-    public function __construct($body = null, array $recipients = [], $sender = null)
+    public function __construct($body, array $recipients, $sender = null)
     {
-        $this
-            ->setBody($body)
-            ->setRecipients($recipients)
-            ->setSender($sender);
-    }
-
-    /**
-     * Sets the message body.
-     *
-     * @param string $body
-     *
-     * @return $this
-     */
-    public function setBody($body)
-    {
-        $this->body = $body;
-
-        return $this;
+        $this->setBody($body);
+        $this->setRecipients($recipients);
+        $this->setSender($sender);
     }
 
     /**
@@ -75,52 +64,13 @@ class Message
     }
 
     /**
-     * Returns the length of the message body in characters.
+     * Returns the length of the message body in a number of characters.
      *
      * @return integer
      */
     public function getLength()
     {
         return mb_strlen($this->getBody(), 'UTF-8');
-    }
-
-    /**
-     * Adds recipient.
-     *
-     * @param string $recipient A phone number in the one of formats: "+71234567890" or "71234567890".
-     *
-     * @return $this
-     *
-     * @throws \InvalidArgumentException If $recipient is not a string or format is invalid.
-     */
-    public function addRecipient($recipient)
-    {
-        array_push(
-            $this->recipients,
-            $this->canonicalizeRecipient($recipient)
-        );
-
-        return $this;
-    }
-
-    /**
-     * Sets/replaces list of recipients.
-     *
-     * @param string[] $recipients A list of phone numbers in the one of formats: "+71234567890" or "71234567890".
-     *
-     * @return $this
-     *
-     * @throws \InvalidArgumentException If one of recipients is not a string or format is invalid.
-     */
-    public function setRecipients(array $recipients)
-    {
-        $this->recipients = [];
-
-        foreach ($recipients as $recipient) {
-            $this->addRecipient($recipient);
-        }
-
-        return $this;
     }
 
     /**
@@ -134,20 +84,6 @@ class Message
     }
 
     /**
-     * Sets the name of sender.
-     *
-     * @param string $sender
-     *
-     * @return $this
-     */
-    public function setSender($sender)
-    {
-        $this->sender = $sender;
-
-        return $this;
-    }
-
-    /**
      * Returns the name of sender.
      *
      * @return string|null
@@ -157,30 +93,104 @@ class Message
         return $this->sender;
     }
 
+    /*
+     * @param string $body
+     */
+    private function setBody($body)
+    {
+        if (is_null($body)) {
+            throw new InvalidArgumentException('Body of message cannot be null');
+        }
+
+        if (!is_string($body)) {
+            throw new InvalidArgumentException(
+                sprintf(
+                    'Body of message must be a string, but "%s" given',
+                    is_object($body) ? get_class($body) : gettype($body)
+                )
+            );
+        }
+
+        if (mb_strlen($body, 'UTF-8') === 0) {
+            throw new InvalidArgumentException('Body of message cannot be an empty string');
+        }
+
+        $this->body = (string) $body;
+    }
+
     /**
-     * @param string $recipient
+     * Sets/replaces list of recipients.
      *
-     * @return string
+     * @param string[] $recipients A list of phone numbers in the one of formats: "+71234567890" or "71234567890".
+     *
+     * @throws \InvalidArgumentException If one of recipients is not a string or format is invalid.
+     */
+    private function setRecipients(array $recipients)
+    {
+        if (count($recipients) === 0) {
+            throw new InvalidArgumentException('At least one recipient required');
+        }
+
+        $this->recipients = [];
+        foreach ($recipients as $recipient) {
+            $this->addRecipient($recipient);
+        }
+    }
+
+    /**
+     * @param string $recipient A phone number in the one of formats: "+71234567890" or "71234567890".
      *
      * @throws \InvalidArgumentException If $recipient is not a string or format is invalid.
      */
-    private function canonicalizeRecipient($recipient)
+    private function addRecipient($recipient)
     {
+        if (is_null($recipient)) {
+            throw new InvalidArgumentException('Recipient cannot be NULL');
+        }
+
         if (!is_string($recipient)) {
-            throw new \InvalidArgumentException(
+            throw new InvalidArgumentException(
                 sprintf(
-                    'Recipient should be a string, but "%s" given.',
+                    'Recipient must be presented as string, but "%s" given',
                     is_object($recipient) ? get_class($recipient) : gettype($recipient)
                 )
             );
         }
 
-        if (preg_match('/^\+?(\d+)$/Ds', $recipient, $matches) !== 1) {
-            throw new \InvalidArgumentException(
-                sprintf('Invalid format of recipient. Should be a string like "+71234567890" or "71234567890"')
+        if (\preg_match('/^\+\d+$/Ds', $recipient) != 1) {
+            throw new InvalidArgumentException(
+                sprintf(
+                    'Recipient must match pattern \+\d+, but "%s" does not match',
+                    $recipient
+                )
             );
         }
 
-        return "+${matches[1]}";
+        array_push($this->recipients, $recipient);
+    }
+
+    /**
+     * @param string $sender
+     */
+    private function setSender($sender)
+    {
+        if (is_null($sender)) {
+            return;
+        }
+
+        if (!is_string($sender)) {
+            throw new InvalidArgumentException(
+                sprintf(
+                    'Sender must be presented as string, but "%s" given',
+                    is_object($sender) ? get_class($sender) : gettype($sender)
+                )
+            );
+        }
+
+        if (mb_strlen(trim($sender), 'UTF-8') === 0) {
+            throw new InvalidArgumentException('Sender cannot be an empty string or contains whitespaces only');
+        }
+
+        $this->sender = $sender;
     }
 }