+++ src/MessageServiceInterface.php
@@ -2,7 +2,7 @@
 /**
  * This file is part of the "litgroup/sms" package.
  *
- * (c) LitGroup <http://litgroup.ru/>
+ * (c) Roman Shamritskiy <roman@litgroup.ru>
  *
  * This source file is subject to the MIT license that is bundled
  * with this source code in the file LICENSE.
@@ -20,13 +20,6 @@ use LitGroup\Sms\Logger\MessageLoggerInterface;
  */
 interface MessageServiceInterface
 {
-    /**
-     * Returns new instance of Message.
-     *
-     * @return Message
-     */
-    public function createMessage();
-
     /**
      * Sends single message.
      *